[Imp] DMF Loader: Offset without note should now work. [Ref] Got rid of some compiler warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@1992 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/load_j2b.cpp
@@ -801,7 +801,7 @@ bool CSoundFile::ReadAM(FileReader &file, ModLoadingFlags loadFlags)
 			if(pan >= 128)
 				ChnSettings[nChn].dwFlags = CHN_MUTE;
 			else
-				ChnSettings[nChn].nPan = std::min(pan * 4, 256);
+				ChnSettings[nChn].nPan = static_cast<uint16>(std::min(pan * 4, 256));
 		}
 	}
